+++ src/main/java/com/prgrms/rg/domain/ridingpost/application/RidingPostReadService.java
@@ -1,0 +1,7 @@
+package com.prgrms.rg.domain.ridingpost.application;
+
+import com.prgrms.rg.domain.ridingpost.model.RidingPost;
+
+public interface RidingPostReadService {
+	RidingPost getRidingPostEntityById(Long postId);
+}
